requirements.txt
amqp==5.3.1
billiard==4.2.4
celery==5.6.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
exceptiongroup==1.3.1
kombu==5.6.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.2.14
